Fix keyboard navigation issues with markdown editors (#17874)
* fix/Keyboard navigation issues with Markdown Editors

Fixes #17854
Filtered out hidden elements, but then I had to filter the toolbar on the additional instances of the component for consistency.
I don't like the look of duplicated code but not much I can do to dry this up.

* fix/Keyboard navigation issues with Markdown Editors

Fixes #17854
Filtered out hidden elements, and had to refactor a little.
I don't like the look of duplicated code but not much I can do to dry this up.